requirements.txt
APScheduler==3.6.3
binance-futures==1.1.0
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
pytz==2020.1
requests==2.23.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.9
websocket-client==0.57.0
pyotp
tweepy
pprint

